refactor(charts): centralize legacy chart parse in ooxml-common parse_chart_part (pptx switched)
Move the legacy DrawingML chart (`c:` namespace) structure parse out of the
pptx `parse_legacy_chart` body and into a shared
`ooxml_common::chart::parse_chart_part(chart_root, color_resolver)` that returns
the crate-owned `ChartModel`. pptx now delegates the whole parse and only keeps
the graphic-frame wrapper (`ChartElement` x/y/w/h). chartEx (`cx:` namespace)
stays pptx-local (PR3), and xlsx is untouched (PR2).

This is a pure, mechanical move. The only edits to the transplanted body are:
- `parse_color_node(fill, theme)` -> `color_resolver.resolve_solid_fill(fill)`
- the `ooxml_common::chart::` self-prefix dropped (now inside that module)
- the local `PptxColorResolver` replaced by the passed `&dyn ColorResolver`
- the `ChartElement` frame wrapper replaced by a bare `ChartModel` return
- `theme.get("+mj-lt"|"+mn-lt")` (theme major/minor Latin faces) now flow
  through two new default `ColorResolver` methods so ooxml-common needs no
  theme-map access; `PptxColorResolver` overrides them.

Supporting additions in ooxml-common: a private `attr` helper identical to the
per-crate one (matches only no-namespace attributes — the shape every chart
attribute uses) so the moved body stays byte-identical, and the two
theme-font accessors on `ColorResolver` (default `None`).

Byte-stability (absolute requirement) verified: a golden harness parsed 14
fixtures covering every branch (bar/line/pie/doughnut/area/scatter/bubble/radar,
combo + secondary axis, dateAx, multi-level cat, str/num literals, deleted axes,
manualLayout, chartSpace border, date1904, empty/absent series) and serialized
each `ChartModel` to JSON; the output is SHA-identical before and after the move.
All existing pptx legacy-chart tests, the core chart renderer-signature snapshot,
and the pptx chart-nested-emit test remain green.
